requirements.txt
pydantic==2.7.4
load-dotenv
streamlit==1.36.0
pdfconverse==0.1.3
google-cloud-firestore==2.10.0
altair==4.1.0
firebase-admin==6.0.1 
google-generativeai==0.7.1
google-cloud-storage==2.17
fastapi
psycopg2-binary==2.9.9